requirements.txt
spacy[en_core_web_sm]==3.5.0
fastapi==0.86.0
python-multipart==0.0.5
torchvision==0.14.0
transformers==4.28.1
uvicorn==0.19.0
